Refactored load() to use require_module() correctly!
For some reason, after reading the Module::Runtime doc I thought
it was going to return false if it failed to load the module,
where in fact it croaks. Discovered this by running Devel::Cover,
and realising that my croak wasn't ever getting called, even though
I thought my test confirmed that it was.